Use SQL BigInteger type to store BGP LOCAL_PREF
The new route control API extension allows to define BGP LOCAL_PREF for
the routes that will be advertised. BGP LOCAL_PREF maximum value is
2^32-1 and the SQL Integer type is limited to 2^31-1.
Cannot add a unit test as unit test framework uses an in-memory sqlite3
DB and its Integer type maximum value is 2^64-1. Need to add a
functional test.

Partial-Bug: #1733905

+++ networking_bgpvpn/neutron/db/bgpvpn_db.py
@@ -111,7 +111,7 @@ class BGPVPNPortAssociationRoute(model_base.BASEV2, model_base.HasId):
     type = sa.Column(sa.Enum(*bgpvpn_routes_control.ROUTE_TYPES,
                              name="bgpvpn_port_assoc_route_type"),
                      nullable=False)
-    local_pref = sa.Column(sa.Integer(),
+    local_pref = sa.Column(sa.BigInteger(),
                            nullable=True)
     # prefix is NULL unless type is 'prefix'
     prefix = sa.Column(sa.String(49),
